Design for command-line functionality.
Support for command-line functionality is to be
added. A draft of the design is described in the
helpfile.

This includes what mechanisms are available, with
explanations of the behaviour of each mechanism.

One mechanism is assigned as the default behaviour
of the command-line functionality.
And other mechanism can be activated through the
use of switches.

It also depicts the layouts for passing parameters
to the program.